+++ example/zip_future_example.dart
@@ -9,6 +9,8 @@ void main() {
 
   zipFuture.execute().then((results) {
     print(results); // [1, 2, 3]
+  }).catchError((error) {
+    print('Error: $error');
   });
 
   // Map the results to sum all elements
@@ -16,5 +18,37 @@ void main() {
     return results.fold(0, (previousValue, element) => previousValue + element);
   }).then((result) {
     print(result); // 6
+  }).catchError((error) {
+    print('Error: $error');
+  });
+
+  // Example with error handling
+  final future4 = Future.value(1);
+  final future5 = Future.error('An error occurred');
+  final future6 = Future.value(3);
+
+  final zipFutureWithError = ZipFuture.zip([future4, future5, future6]);
+
+  zipFutureWithError.execute(onError: (index, error) {
+    print('Error at index $index: $error');
+  }).then((results) {
+    print(results); // [1, 3]
+  }).catchError((error) {
+    print('Unhandled Error: $error');
+  });
+
+  // Map the results to sum all elements with error handling
+  zipFutureWithError.executeThenMap<num>(
+    (results) {
+      return results.fold(
+          0, (previousValue, element) => previousValue + element);
+    },
+    onError: (index, error) {
+      print('Error at index $index: $error');
+    },
+  ).then((result) {
+    print(result); // 4
+  }).catchError((error) {
+    print('Unhandled Error: $error');
   });
 }