+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.1.0
+
+- Added error handling with optional onError callback for execute and executeThenMap methods.
+- Updated README with examples demonstrating error handling.
+- Added dart docs.